Updated Exercise step calculation
Previoulsy if all the activities have heart rate below the aerobic zone then
we were taking the steps of the activity with smallest duration as exercise steps.
Now we have dropped this logic. We leave it up to the user to decide whether
to include the steps for any particular activity in "exercise steps" or not.