Updated MCs calculation to improve MCs score
Earlier if there is no step data for any hour, we classify that hour as
"inactive". So for example if it is 8:30 AM EST then obviously there
would be no step data for any time after 8:30 AM (future data!). So
in that case all the hours 9:00 am - 9:59 am, 10:00 am - 10:59 am etc are
classified as "inactive" hours because there is no step data present yet.
Because of this, "inactive" hours count increases, resulting in bad MCs score.
Ideally we should not count those future hours in inactive hours.

To solve this problem, we are determing the local time of user at
the time of calculation and marking any hour after the current local
time of user as "no data yet" if there is no steps data.