Refactored code to use "get_filtered_activity_stats" method
In the list of activities which were provided to the user input form, the activities
which were manually edited were not having steps information. A function
"get_filtered_activity_stats" was already written to deal such cases, but here
it was not used, instead whole calculation was rewritten but missing the piece
which handles no step case. So the code is refactored to use this function
to maintain consistency. Also it resolved the issue mentioned.

Resolves - #160372105 (https://www.pivotaltracker.com/story/show/160372105)