Fixed issue where capping was not done per quarter
Earlier if a quarter (eg 9:00:00 - 9:14:59) have 3 epochs. Then if active minute
for each epoch is more than 15 min then we are capping to 15 min. But actually this
was wrong way. Capping should be done on total of active seconds for all 3 epochs.

So if total active minutes result in 30 minutes then it should be capped to 15 minute