Updated sleep average calculation
Updated sleep average calculation so that the denominator is the actual number of days sleep is
reported by the user (either by userinput or syncing watch) within the range selected.
For example, if user want to see the average of 1 week, where only for 3 days sleep information
is present, then for average for 7 days, denominator will be 3 instead of 7.

Ticket - #161572443 (https://www.pivotaltracker.com/story/show/161572443)