raw-cal: Handle the case when activity duration is not valid.
Handle the case when activity duration is not valid duration
(ie. not in seconds). This usually happens when the user edit
or create new activity and do not provide correct activity
start or end date.

TODO: Add frontend validation to show error message if
user missed start or end date of the activity.

Fixes: #164189667 (https://www.pivotaltracker.com/story/show/164189667)