raw-data: Update steps calculation wrt manual steps.
This fixes an issue which adds the steps from manual
activity to the total steps. Complete manual steps
or some part of those steps are already part of
total daily steps and were double counted and added
to the total steps.

There are 3 different cases which are documented
here - https://www.pivotaltracker.com/story/show/164454853
(check the comments)

Fixes: #164454853