MCS: Change priority of the "exercise".
Earlier, if for any hour user is having both
sleep and exercise then "sleeping" was given
more priority and hour is marked as "sleeping".

Now we have changed the priority and in case
of such conflict, "exercise" will be given
more priority than "sleeping". This is
because the user would like to see "exercise"
more than "sleeping".